{"schema":"decodex/commit/1","summary":"Split docs and benchmark tests into focused modules","authority":"manual"}
